Use configuration instead of pure discovery for XS Support
Given in Gentoo that ::XS is either going to be available, or not,
based on user configuration, *before* configure runs, trying to turn
it on/off in configure is nonsense.

Subsequently, this test logic just determines if we should be running
the XS tests or not.

If ::XS is installed already, then XS tests should be run despite user
not opting in to XS, as the opt-in only regulates dependencies, but
LMU can still trigger ::XS at runtime.

If the user opts in to XS, then the tests should be run regardless of
perl detecting XS being installed, because XS *NOT* being available
is then an error.